Remove examples/
From this point, I will maintain this gh-pages branch different from the
master branch.

The master is to be used for syncing with reveal.js and making
contribution in need in future.

The gh-pages branch is to be used for creating and storing my slides.

I will manually check updates from master to gh-pages.